fix: validate CLI and Obsidian release artifacts
