+++ packages/cli/src/index.ts
@@ -3,7 +3,7 @@
 import { NodeRuntime } from "@effect/platform-node";
 import chalk from "chalk";
 import boxen from "boxen";
-import { Cause, Console, Effect } from "effect";
+import { Console, Effect } from "effect";
 import {
 	ConfluenceSettingsLive,
 	ConfluenceUploadSettings,
@@ -15,6 +15,7 @@ import {
 } from "@markdown-confluence/lib";
 import { PuppeteerMermaidRenderer } from "@markdown-confluence/mermaid-puppeteer-renderer";
 import { ConfluenceClient } from "confluence.js";
+import { getErrorMessage } from "./errorMessage";
 
 const program = Effect.gen(function* () {
 	const runtimeEnvironment = yield* RuntimeEnvironmentService as any;
@@ -82,77 +83,3 @@ NodeRuntime.runMain(
 		Effect.provide(MarkdownConfluencePlatformLive),
 	) as Effect.Effect<void, never>,
 );
-
-function getErrorMessage(error: unknown): string {
-	const message = formatUnknownError(error, new WeakSet());
-	return message.trim().length > 0 ? message : "Unknown error";
-}
-
-function formatUnknownError(error: unknown, seenErrors: WeakSet<object>): string {
-	if (Cause.isCause(error)) {
-		return Cause.pretty(error);
-	}
-
-	if (error instanceof Error) {
-		const errorMessage = error.message.trim();
-		const message = errorMessage.length > 0 ? errorMessage : error.stack || error.name;
-		const causeMessage = formatErrorCause(error.cause, seenErrors);
-		return joinMessages([message, causeMessage]);
-	}
-
-	if (typeof error === "string") {
-		return error;
-	}
-
-	if (error === null) {
-		return "null";
-	}
-
-	if (error === undefined) {
-		return "undefined";
-	}
-
-	if (typeof error === "object") {
-		if (seenErrors.has(error)) {
-			return "[Circular error]";
-		}
-		seenErrors.add(error);
-
-		const message = getStringProperty(error, "message");
-		const causeMessage = formatErrorCause(getObjectProperty(error, "cause"), seenErrors);
-		const serializedError = stringifyObject(error);
-		return joinMessages([message, serializedError, causeMessage]);
-	}
-
-	return String(error);
-}
-
-function formatErrorCause(cause: unknown, seenErrors: WeakSet<object>): string {
-	if (cause === undefined) {
-		return "";
-	}
-
-	return `Caused by: ${formatUnknownError(cause, seenErrors)}`;
-}
-
-function getObjectProperty(source: object, propertyName: string): unknown {
-	return propertyName in source ? (source as Record<string, unknown>)[propertyName] : undefined;
-}
-
-function getStringProperty(source: object, propertyName: string): string {
-	const propertyValue = getObjectProperty(source, propertyName);
-	return typeof propertyValue === "string" ? propertyValue : "";
-}
-
-function stringifyObject(source: object): string {
-	try {
-		const serializedError = JSON.stringify(source, null, 2);
-		return serializedError && serializedError !== "{}" ? serializedError : "";
-	} catch {
-		return "";
-	}
-}
-
-function joinMessages(messages: string[]): string {
-	return messages.filter((message) => message.trim().length > 0).join("\n");
-}
