+++ vite.package-build.ts
@@ -11,6 +11,9 @@ export const generatedBanner = `/*
 THIS IS A GENERATED/BUNDLED FILE BY VITE
 if you want to view the source, please visit the github repository of this plugin
 */
+`;
+
+export const generatedNodeBanner = `${generatedBanner}
 import { createRequire as __createRequire } from "node:module";
 const require = __createRequire(import.meta.url);
 `;
